test(network): move get_connected_stream_futures to other file
