+++ crates/papyrus_network/src/streamed_data_protocol/protocol_test.rs
@@ -1,7 +1,4 @@
-use futures::{AsyncRead, AsyncWrite, AsyncWriteExt, Future, StreamExt};
-use libp2p::core::multiaddr::multiaddr;
-use libp2p::core::transport::memory::MemoryTransport;
-use libp2p::core::transport::{ListenerId, Transport};
+use futures::AsyncWriteExt;
 use libp2p::core::upgrade::{write_varint, InboundUpgrade, OutboundUpgrade};
 use libp2p::core::UpgradeInfo;
 use pretty_assertions::assert_eq;
@@ -11,6 +8,7 @@ use crate::messages::block::{BlockHeader, GetBlocks, GetBlocksResponse};
 use crate::messages::common::{BlockId, Fin};
 use crate::messages::proto::p2p::proto::get_blocks_response::Response;
 use crate::messages::{read_message, write_message};
+use crate::test_utils::get_connected_stream_futures;
 
 fn hardcoded_responses() -> Vec<GetBlocksResponse> {
     vec![
@@ -44,29 +42,6 @@ fn both_protocols_have_same_info() {
         outbound_protocol.protocol_info().collect::<Vec<_>>(),
         inbound_protocol.protocol_info().collect::<Vec<_>>()
     );
-}
-
-async fn get_connected_stream_futures() -> (
-    impl Future<Output = impl AsyncRead + AsyncWrite>,
-    impl Future<Output = impl AsyncRead + AsyncWrite>,
-) {
-    let address = multiaddr![Memory(0u64)];
-    let mut transport = MemoryTransport::new().boxed();
-    transport.listen_on(ListenerId::next(), address).unwrap();
-    let listener_addr = transport
-        .select_next_some()
-        .await
-        .into_new_address()
-        .expect("MemoryTransport not listening on an address!");
-
-    (
-        async move {
-            let transport_event = transport.next().await.unwrap();
-            let (listener_upgrade, _) = transport_event.into_incoming().unwrap();
-            listener_upgrade.await.unwrap()
-        },
-        async move { MemoryTransport::new().dial(listener_addr).unwrap().await.unwrap() },
-    )
 }
 
 #[tokio::test]
